pulizia repo;functions moved into modules folder
